[WIP] Add open-source libsensors
IIO hal based on Motion Apps 5.1.6 Release Candidate 13, the closest one
I could find to match with our prebuilt HAL (which is 5.1.5).
The different parts were carefully compared with the disassembly to make
them work equally. libmllite.so and libmplmpu.so contains the mathematic
functions, and are untouched; however, it's possible to open-source them
later on as well.

Light sensor HAL needs kernel commit f4503b3, which is already merged.

Change-Id: I3d5a9b84d868b73573425948cc2339df60249340