Update patch set 2
Patch Set 2:

I was going to reply on the abandoned commit, but I'm not sure parts of this are needed. Aren't the F2FS tools included automatically since MM (in target/product/core_minimal.mk)? Also, can we get rid of the setup_fs service entirely? It fully clicked for me now that setup_fs is the source of our F2FS cache issues. I removed it on my devices since LP (since we're wiping everything with the first clean install), and had forgotten about it.

Patch-set: 2